Move mobile video serialization logic to serializers.py
